+++ package/thunderstore/Waystones/CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.0.1
+* config options to disable directions
+* items from configurable json/yaml list can be sacrificed to waystone to reduce cooldown
+
 # 1.0.0
 * Initial release